Freeze graph and test it
Freeze the graph of evaliation with model with selected weights.
During training weights are saved but not the relationship between
weights. These saved weight need to be restored to model which describes
this relationship and freeze to `.pb` file.

Frozen file is tested by loading to a new graph and passing an input
image into it.